IsGithubNixpkgsURL and HashFromNixPkgsURL to nix
