+++ internal/nix/nixpkgs.go
@@ -11,10 +11,11 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
+	"regexp"
+	"strings"
 
 	"github.com/fatih/color"
 	"github.com/pkg/errors"
-
 	"go.jetpack.io/devbox/internal/fileutil"
 	"go.jetpack.io/devbox/internal/xdg"
 )
@@ -109,4 +110,26 @@ func saveToNixpkgsCommitFile(commit string, commitToLocation map[string]string) 
 func nixpkgsCommitFilePath() string {
 	cacheDir := xdg.CacheSubpath("devbox")
 	return filepath.Join(cacheDir, "nixpkgs.json")
+}
+
+// IsGithubNixpkgsURL returns true if the package is a flake of the form:
+// github:NixOS/nixpkgs/...
+//
+// While there are many ways to specify this input, devbox always uses
+// github:NixOS/nixpkgs/<hash> as the URL. If the user wishes to reference nixpkgs
+// themselves, this function may not return true.
+func IsGithubNixpkgsURL(url string) bool {
+	return strings.HasPrefix(url, "github:NixOS/nixpkgs/")
+}
+
+var hashFromNixPkgsRegex = regexp.MustCompile(`github:NixOS/nixpkgs/([^#]+).*`)
+
+// HashFromNixPkgsURL will (for example) return 5233fd2ba76a3accb5aaa999c00509a11fd0793c
+// from github:nixos/nixpkgs/5233fd2ba76a3accb5aaa999c00509a11fd0793c#hello
+func HashFromNixPkgsURL(url string) string {
+	matches := hashFromNixPkgsRegex.FindStringSubmatch(url)
+	if len(matches) == 2 {
+		return matches[1]
+	}
+	return ""
 }
